renaming joystick to gamepad
