+++ libs/joystick/docs/reference/joystick/set-button.md
@@ -1,7 +1,7 @@
 # set Button
 
 ```sig
-gamepad.setButton(0, false)
+joystick.setButton(0, false)
 ```
 
 Sets the button state to up (true) or down (false)
