+++ libs/joystick/joystick.ts
@@ -1,4 +1,4 @@
 //% icon="\uf11b" color="#777777"
-namespace gamepad {
+namespace joystick {
     
 }
