+++ libs/joystick/sim/state.ts
@@ -1,0 +1,2 @@
+namespace pxsim {
+}
