+++ pxtarget.json
@@ -17,7 +17,7 @@
         "libs/pulse",
         "libs/cable",
         "libs/infrared",
-        "libs/gamepad",
+        "libs/joystick",
         "libs/tests",
         "libs/behaviors"
     ]
